refactored common model files
