+++ src/ResourceManager/RecoveryServices.Backup/Commands.RecoveryServices.Backup.Models/Commands.RecoveryServices.Backup.Models.csproj
@@ -41,9 +41,10 @@
   <ItemGroup>
     <Compile Include="BaseObjects.cs" />
     <Compile Include="CmdletParamEnums.cs" />
-    <Compile Include="CommonEnums.cs" />
-    <Compile Include="CommonPolicyRetentionObjects.cs" />
-    <Compile Include="CommonPolicyScheduleObjects.cs" />
+    <Compile Include="CommonModels\CommonUtils.cs" />
+    <Compile Include="CommonModels\CommonEnums.cs" />
+    <Compile Include="CommonModels\CommonPolicyRetentionObjects.cs" />
+    <Compile Include="CommonModels\CommonPolicyScheduleObjects.cs" />
     <Compile Include="IaasVmModels\AzureRmRecoveryServicesIaasVmPolicy.cs" />
     <Compile Include="IaasVmModels\AzureRmRecoveryServicesIaasVmItem.cs" />
     <Compile Include="IaasVmModels\AzureRmRecoveryServicesIaasVmContainer.cs" />
